requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.4
django-admin-interface==0.15.5
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-tinymce==3.2.0
gunicorn==20.1.0
idna==2.10
Pillow==8.1.0
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
